move ZoomingTool component out of Geral subfolder
