+++ website/src/Components/ZoomingTool/ZoomingTool.js
